+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 0.17.3
 ------
 - Fix duplicates when using custom through association class on M2M relations
+- Fix `update_or_create`. (#721)
 
 0.17.2
 ------